Summary
Overview
This project is combines the idea of IOT and smart home in one useful
project.
The project aims to save you the time that you take to check your post
mail box daily and simply send you notification on your phone if you
have a new post in your mailbox.
One of the features that is offered in that project is the connection
coverage since your post box will be connected to internet ; which
allows the user to check his mail box from any place in the world
freely.
Goals
Detect if there is a new post inside the postbox.
Notify the user by the mobile app with the detection result.